requirements-base.txt
environs
requests
pytz
structlog
piexif
filetype
pillow
requests_futures
requests-toolbelt
python-dateutil